requirements.txt
wcwidth==0.2.13
typing-extensions==4.7.1
traitlets==5.9.0
sqlparse==0.4.4
six==1.16.0
setuptools==39.0.1
pytz==2024.1
pygments==2.17.2
psycopg2==2.9.9
prompt-toolkit==3.0.47
pip==24.0
pickleshare==0.7.5
parso==0.8.4
matplotlib-inline==0.1.6
jedi==0.19.1
ipython==7.34.0
Django==3.2.23
django-extensions==3.2.3
django-extensions-shell==1.7.4.1
django-debug-toolbar==3.3.0
dj-database-url==2.2.0
decorator==5.1.1
colorama==0.4.6
backcall==0.2.0
asgiref==3.7.2
gunicorn==23.0.0
